+++ src/constants/formFields.js
@@ -1,7 +1,7 @@
 import $select from "../components/common/select.js";
 import $input from "../components/common/input.js";
 import $textarea from "../components/common/textarea.js";
-import { categoryOptions, distanceOptions } from "../data/selectOptions.js";
+import { categoryOptions, distanceOptions } from "./selectOptions.js";
 import { validateRestaurantForm } from "../validation/validationForm.js";
 
 const senseChangeRestaurantFormValue = () => {
